arch-mvi-21 Improved classes structure + test coverage
