+++ mvi-sideeffect-solo/src/test/kotlin/dev/sunnyday/arch/mvi/side_effect/solo/filter/side_effect/SideEffectsWithTypeInstanceFilterTest.kt
@@ -1,9 +1,7 @@
 package dev.sunnyday.arch.mvi.side_effect.solo.filter.side_effect
 
-import dev.sunnyday.arch.mvi.primitive.ObservableEvent
 import dev.sunnyday.arch.mvi.side_effect.solo.ExecutingSideEffect
 import dev.sunnyday.arch.mvi.side_effect.solo.SoloSideEffect
-import dev.sunnyday.arch.mvi.side_effect.solo.SoloSideEffectHandler
 import dev.sunnyday.arch.mvi.side_effect.solo.test.TestFilterExecutingSideEffect
 import kotlinx.coroutines.flow.Flow
 import org.junit.jupiter.api.Assertions.*
@@ -104,12 +102,15 @@ class SideEffectsWithTypeInstanceFilterTest {
         )
     }
 
+    @Suppress("EqualsOrHashCode")
     private open class Super {
+
         override fun equals(other: Any?): Boolean {
             return other is Super
         }
     }
 
+    @Suppress("EqualsOrHashCode")
     private class Nested : Super() {
         override fun equals(other: Any?): Boolean {
             return other is Nested
